fixed modal include

+++ application/controllers/Position.php
@@ -62,11 +62,19 @@ class Position extends CI_Controller
 
     public function delete()
     {
-        if (!empty($_POST)) {
-            $this->position_model->delete($_POST['position_id']);
+        if (!empty($_POST))
+        {
+            $res = $this->position_model->get_references($_POST['position_id']);
+            if (empty($res))
+            {
+                $this->position_model->delete($_POST['position_id']);
+            }
+            else
+            {
+                $error_msg = "<h6 id='error' style='color: rgba(255,0,0,0.7)' hidden>Selected position is deleted unsuccessfully. The selected position is currently being used.</h6>";
+                echo $error_msg;
+            }
         }
         redirect(base_url('position'));
     }
-
-
 }
